# services/payment-service/openapi.yaml
openapi: "3.0.3"
info:
  title: AeroLink Payment Service
  description: PCI-DSS compliant payment processing stub. Full card numbers are never stored or logged.
  version: "1.0.0"
servers:
  - url: https://api.aerolink.app/payments
  - url: http://localhost:3004

tags:
  - name: payments
    description: Payment processing and history
  - name: refunds
    description: Refund operations

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      required: [statusCode, message, correlationId]
      properties:
        statusCode: { type: integer }
        message: { type: string }
        correlationId: { type: string, format: uuid }

    TransactionStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED, PARTIALLY_REFUNDED]

    TransactionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        passengerId: { type: string, format: uuid }
        amount:
          type: object
          properties:
            amount: { type: number, example: 459.99 }
            currency: { type: string, example: "GBP" }
        status: { $ref: "#/components/schemas/TransactionStatus" }
        paymentProvider: { type: string, example: "STRIPE_STUB" }
        cardLastFour: { type: string, example: "4242", description: "Last 4 digits only — PCI-DSS compliance" }
        cardBrand: { type: string, example: "VISA" }
        processedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }

    RefundRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [CUSTOMER_REQUESTED, FLIGHT_CANCELLED, DUPLICATE_PAYMENT, FRAUD]
        amount:
          type: object
          nullable: true
          description: If null, full refund is issued
          properties:
            amount: { type: number }
            currency: { type: string }

    RefundResponse:
      type: object
      properties:
        refundId: { type: string, format: uuid }
        transactionId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        refundAmount:
          type: object
          properties:
            amount: { type: number }
            currency: { type: string }
        reason: { type: string }
        status: { type: string, enum: [PENDING, COMPLETED, FAILED] }
        estimatedSettlement: { type: string, format: date, description: "3-5 business days" }
        createdAt: { type: string, format: date-time }

    PaginatedTransactions:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/TransactionResponse" } }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

paths:
  /payments:
    get:
      tags: [payments]
      summary: List transactions (AIRLINE_ADMIN only; CUSTOMER sees own via booking context)
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TransactionStatus" }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedTransactions" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /payments/{id}:
    get:
      tags: [payments]
      summary: Get transaction detail
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionResponse" }
        "403":
          description: Forbidden — not your transaction
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /payments/{id}/refund:
    post:
      tags: [refunds]
      summary: Issue a refund (AIRLINE_ADMIN only)
      operationId: issueRefund
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefundRequest" }
      responses:
        "202":
          description: Refund initiated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefundResponse" }
        "409":
          description: Transaction already fully refunded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Cannot refund a FAILED transaction
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /health:
    get:
      tags: [payments]
      summary: Liveness probe
      operationId: health
      responses:
        "200": { description: Healthy }

  /ready:
    get:
      tags: [payments]
      summary: Readiness probe
      operationId: ready
      responses:
        "200": { description: Ready }
        "503": { description: Not ready }